+++ bc/settings/base.py
@@ -79,6 +79,7 @@ INSTALLED_APPS = [
     "bc.recruitment_api",
     "bc.search.apps.SearchConfig",
     "bc.standardpages",
+    "bc.step_by_step",
     "bc.users",
     "bc.utils",
     "wagtail_transfer",
@@ -86,13 +87,13 @@ INSTALLED_APPS = [
     "wagtailorderable",
     "wagtail_automatic_redirects",
     "wagtail.contrib.modeladmin",
-    "wagtail.contrib.postgres_search",
     "wagtail.contrib.settings",
     "wagtail.contrib.search_promotions",
     "wagtail.contrib.forms",
     "wagtail.contrib.redirects",
     "wagtail.contrib.routable_page",
     "wagtail.contrib.table_block",
+    "wagtail.contrib.typed_table_block",
     "wagtail.contrib.legacy.richtext",
     "wagtail.embeds",
     "wagtail.sites",
@@ -240,7 +241,7 @@ if "BONSAI_URL" in env:
     }
 else:
     WAGTAILSEARCH_BACKENDS = {
-        "default": {"BACKEND": "wagtail.contrib.postgres_search.backend"}
+        "default": {"BACKEND": "wagtail.search.backends.database"}
     }
 # Reduction factor between 0 and 1 to apply to the relevanve score of search
 # results with the NewsPage content type. See bc.search.elasticsearch5.
@@ -748,6 +749,11 @@ WAGTAILDOCS_DOCUMENT_MODEL = "documents.CustomDocument"
 PASSWORD_REQUIRED_TEMPLATE = "patterns/pages/wagtail/password_required.html"
 
 
+# Default field for automatic primary keys. (Introduced in Django 3.2)
+# https://docs.djangoproject.com/en/3.2/ref/settings/#std:setting-DEFAULT_AUTO_FIELD
+DEFAULT_AUTO_FIELD = "django.db.models.AutoField"
+
+
 # Default size of the pagination used on the front-end.
 DEFAULT_PER_PAGE = 10
 
@@ -805,9 +811,6 @@ RESPOND_COMPLIMENTS_WEBSERVICE = env.get("RESPOND_COMPLIMENTS_WEBSERVICE")
 RESPOND_DISCLOSURES_WEBSERVICE = env.get("RESPOND_DISCLOSURES_WEBSERVICE")
 RESPOND_GET_CATEGORIES_WEBSERVICE = env.get("RESPOND_GET_CATEGORIES_WEBSERVICE")
 RESPOND_GET_FIELDS_WEBSERVICE = env.get("RESPOND_GET_FIELDS_WEBSERVICE")
-
-# MapIt.MySociety API credentials
-MAPIT_API_KEY = env.get("MAPIT_API_KEY")
 
 # Google maps API credentials
 GOOGLE_MAPS_V3_APIKEY = env.get("GOOGLE_MAPS_V3_APIKEY")