+++ bc/static_src/javascript/main.js
@@ -18,6 +18,8 @@ import Filters from './components/job-filters';
 import FeedbackWidget from './components/feedback-widget';
 import Chart from './components/chart';
 import GoogleMap from './components/map';
+import Steps from './components/steps';
+import ContentsScroller from './components/contents-scroller';
 
 // Add polyfill fix for forEach carousel
 import foreachPolyfill from './polyfills/foreach-polyfill';
@@ -126,6 +128,16 @@ document.addEventListener('DOMContentLoaded', () => {
 
     for (const googlemap of document.querySelectorAll(GoogleMap.selector())) {
         new GoogleMap(googlemap);
+    }
+
+    for (const stepsContainer of document.querySelectorAll(Steps.selector())) {
+        new Steps(stepsContainer);
+    }
+
+    for (const contentsComponent of document.querySelectorAll(
+        ContentsScroller.selector(),
+    )) {
+        new ContentsScroller(contentsComponent);
     }
 
     // Test react - add a div with a data attribute of `data-test-react` to test