+++ bc/utils/blocks.py
@@ -7,14 +7,27 @@ from django.utils.functional import cached_property
 
 from wagtail.admin.staticfiles import versioned_static
 from wagtail.contrib.table_block.blocks import TableBlock
+from wagtail.contrib.typed_table_block.blocks import TypedTableBlock
 from wagtail.core import blocks
 from wagtail.documents.blocks import DocumentChooserBlock
 from wagtail.embeds.blocks import EmbedBlock
 from wagtail.images.blocks import ImageChooserBlock
 
 from .constants import RICH_TEXT_FEATURES
+from .models import ImportantPages
 from .utils import is_number
 from .widgets import BarChartInput, LineChartInput, PieChartInput
+
+
+class CaptionedTableBlock(blocks.StructBlock):
+    table = TypedTableBlock(
+        [("numeric", blocks.DecimalBlock()), ("rich_text", blocks.RichTextBlock())],
+    )
+    caption = blocks.TextBlock(required=False)
+
+    class Meta:
+        icon = "table"
+        template = "patterns/molecules/streamfield/blocks/captioned_table_block.html"
 
 
 class ImageBlock(blocks.StructBlock):
@@ -62,14 +75,30 @@ class HighlightBlock(blocks.RichTextBlock):
 
 
 class LocalAreaLinksBlock(blocks.StructBlock):
+    heading = blocks.RichTextBlock(
+        features=RICH_TEXT_FEATURES, default="<p><b>Find local information</b></p>",
+    )
     introduction = blocks.RichTextBlock(
         features=RICH_TEXT_FEATURES,
-        default="<p>Select your local area for information:</p>",
+        default=(
+            "<p>While we finish building this new website, we’re keeping some local"
+            " information on our old council websites</p>"
+        ),
     )
-    aylesbury_vale_url = blocks.URLBlock(required=False, label="Aylesbury Vale URL")
-    chiltern_url = blocks.URLBlock(required=False, label="Chiltern URL")
-    south_bucks_url = blocks.URLBlock(required=False, label="South Bucks URL")
-    wycombe_url = blocks.URLBlock(required=False, label="Wycombe URL")
+    postcode_lookup_text = blocks.RichTextBlock(
+        features=RICH_TEXT_FEATURES,
+        default="<p>Enter your postcode to help us redirect you to the right place.</p>",
+        help_text="The text that appears on top of the postcode lookup input",
+    )
+    area_lookup_text = blocks.RichTextBlock(
+        features=RICH_TEXT_FEATURES,
+        default="<p>Select your local area to help us direct you to the right place:</p>",
+        help_text="The text that appears on top of the list of local area links",
+    )
+    aylesbury_vale_url = blocks.URLBlock(label="Aylesbury Vale URL")
+    chiltern_url = blocks.URLBlock(label="Chiltern URL")
+    south_bucks_url = blocks.URLBlock(label="South Bucks URL")
+    wycombe_url = blocks.URLBlock(label="Wycombe URL")
 
     class Meta:
         icon = ""
@@ -77,14 +106,17 @@ class LocalAreaLinksBlock(blocks.StructBlock):
 
     def get_context(self, value, parent_context=None):
         context = super().get_context(value, parent_context=parent_context)
-        context["has_area_links"] = any(
-            [
-                value["aylesbury_vale_url"],
-                value["chiltern_url"],
-                value["south_bucks_url"],
-                value["wycombe_url"],
-            ]
-        )
+        context["link_urls"] = {
+            "Aylesbury Vale": value["aylesbury_vale_url"],
+            "Chiltern": value["chiltern_url"],
+            "South Bucks": value["south_bucks_url"],
+            "Wycombe": value["wycombe_url"],
+        }
+        if parent_context is not None and parent_context.get("request"):
+            request = parent_context["request"]
+            context["contact_us_page"] = ImportantPages.for_request(
+                request
+            ).contact_us_page
         return context
 
 
@@ -391,7 +423,7 @@ class BaseStoryBlock(blocks.StreamBlock):
     image = ImageBlock()
     embed = EmbedBlock()
     local_area_links = LocalAreaLinksBlock()
-    table = TableBlock()
+    table = CaptionedTableBlock()
     button = ButtonBlock()
     highlight = HighlightBlock()
 
@@ -401,6 +433,8 @@ class BaseStoryBlock(blocks.StreamBlock):
 
 
 class NestedStoryBlock(BaseStoryBlock):
+    table = TableBlock()
+
     def __init__(self, local_blocks=None, **kwargs):
         super().__init__(**kwargs)
         # Bump down template for heading fields so headings don't clash with those outside the accordion
@@ -455,6 +489,7 @@ class DetailBlock(blocks.StructBlock):
 class StoryBlock(BaseStoryBlock):
     accordion = Accordion()
     detail = DetailBlock()
+    table = CaptionedTableBlock()
 
 
 class ImageOrEmbedBlock(blocks.StructBlock):